refactor(replication): move config contracts into crate
